test: convert bdd to it
