+++ cypress/accessibility/a11y-second-part.test.js
@@ -1,0 +1,3 @@
+import tests from "../support/accessibility/a11y-utils";
+
+tests(300, 600);
