+++ cypress/features/regression/common/icon.feature
