+++ cypress/locators/action-popover/index.js
@@ -2,6 +2,7 @@ import {
   ACTION_POPOVER_BUTTON,
   ACTION_POPOVER_DATA_COMPONENT,
   ACTION_POPOVER_SUBMENU,
+  ACTION_POPOVER_WRAPPER,
 } from "./locators";
 
 // component preview locators
@@ -13,3 +14,4 @@ export const actionPopoverSubmenu = (index) =>
   cy.get(ACTION_POPOVER_SUBMENU).eq(1).children().eq(index);
 export const actionPopoverSubmenuByIndex = () =>
   cy.get(ACTION_POPOVER_SUBMENU).eq(1);
+export const actionPopoverWrapper = () => cy.get(ACTION_POPOVER_WRAPPER);
