+++ cypress/locators/action-popover/locators.js
@@ -3,3 +3,5 @@ export const ACTION_POPOVER_BUTTON = '[data-component="action-popover-button"]';
 export const ACTION_POPOVER_DATA_COMPONENT =
   '[data-component="action-popover"]';
 export const ACTION_POPOVER_SUBMENU = '[data-element="action-popover-submenu"]';
+export const ACTION_POPOVER_WRAPPER =
+  '[data-component="action-popover-wrapper"]';
