+++ package.json
@@ -118,7 +118,6 @@
     "eslint-plugin-react": "^7.20.3",
     "eslint-plugin-react-hooks": "^4.2.0",
     "events": "~1.1.1",
-    "fs-extra": "^10.0.0",
     "husky": "^4.3.6",
     "jest": "^26.6.3",
     "jest-styled-components": "^6.3.1",
@@ -145,8 +144,6 @@
     "classnames": "~2.2.6",
     "crypto-js": "~3.3.0",
     "dotenv": "^10.0.0",
-    "escape-string-regexp": "^4.0.0",
-    "fs-extra": "^10.0.0",
     "immutable": "~3.8.2",
     "invariant": "^2.2.4",
     "lodash": "^4.17.20",
